Rearrange utils and import order
